requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.9.1
PyJWT==1.7.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.11.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
